Change border; change submit button color
Added a hover state and a gradient to the hover
I wanted to include a transition to the gradient but
it doesn't appear this is possible-- or at least
very easy.

Same goes for line up top.
I initially wanted to have a gradient for
the line color.. but that again appears
A
difficult.